+++ client/src/MidiComponents/MidiNotes.tsx
@@ -40,6 +40,12 @@ function MidiNotes(props: MidiNotesProps) {
 
   useEffect(() => {
     // console.log(props.midiNoteInfo)
+<<<<<<< HEAD
+=======
+    let key: string;
+    let start: number;
+    let end: number;
+>>>>>>> bbc93c9 (feat(client): Added display to show midi time in seconds. fix(client): Various bug fixes.)
     // setWidths({});
     setWidths((widths) => {
       let state = {...widths}
@@ -49,6 +55,7 @@ function MidiNotes(props: MidiNotesProps) {
         }
       });
       props.midiNoteInfo.forEach((noteStart) => {
+<<<<<<< HEAD
         let key = Object.keys(noteStart)[0];
         let start = noteStart[key].keyPressed!.start!;
         let end = noteStart[key].keyPressed!.end!;
@@ -62,6 +69,24 @@ function MidiNotes(props: MidiNotesProps) {
       return state;
     })
   }, [props.pulseNum, props.midiNoteInfo])
+=======
+        key = Object.keys(noteStart)[0];
+        start = noteStart[key].keyPressed!.start!;
+        end = noteStart[key].keyPressed!.end!;
+        // console.warn(noteStart[key].keyPressed)
+
+        if(noteStart[key].keyPressed!.start) {
+          if(end === -1) {
+            state[key] = {start: start, end: end};
+          } else {
+            state[key] = {start: start, end: end};
+          }
+        }
+      })
+      return state;
+    })
+  }, [props.midiNoteInfo])
+>>>>>>> bbc93c9 (feat(client): Added display to show midi time in seconds. fix(client): Various bug fixes.)
 
   useEffect(() => {
     setMidiNotes([])
@@ -70,6 +95,7 @@ function MidiNotes(props: MidiNotesProps) {
       const noteTrackChilds: NoteTrackChilds = {};
       const midiNotesArr: ReactPortal[] = []
       props.midiNoteInfo.forEach((noteStart) => {
+<<<<<<< HEAD
         let key = Object.keys(noteStart)[0];
 
         // console.log('midiNoteInfo', props.midiNoteInfo);
@@ -92,6 +118,25 @@ function MidiNotes(props: MidiNotesProps) {
           noteTrackChilds[noteStart[key].noteTrackId].push(elem)
         }
         
+=======
+        // console.log(widths)
+        let key = Object.keys(noteStart)[0];
+        let left = widths[key].start! / (props.midiLength * props.pulseRate) * props.noteTracksRef.current!.offsetWidth + 1;
+        // let width = (noteStart[key].keyPressed!.end! - noteStart[key].keyPressed!.start!) / (props.midiLength * props.pulseRate) * props.noteTracksRef.current!.offsetWidth;
+        let width = (widths[key].end! - widths[key].start) / (props.midiLength * props.pulseRate) * props.noteTracksRef.current!.offsetWidth;
+        if(widths[key].end === -1) {
+          width = (props.pulseNum - widths[key].start) / (props.midiLength * props.pulseRate) * props.noteTracksRef.current!.offsetWidth;
+        }
+        var elem = createElement('span', {...noteStart[key].props, key: noteStart[key].key, style: {
+          height: '23.5px',
+          left: `${left}px` ,
+          width: `${width}px`,
+        }});
+        if(!noteTrackChilds[noteStart[key].noteTrackId]) {
+          noteTrackChilds[noteStart[key].noteTrackId] = [];
+        }
+        noteTrackChilds[noteStart[key].noteTrackId].push(elem)
+>>>>>>> bbc93c9 (feat(client): Added display to show midi time in seconds. fix(client): Various bug fixes.)
       });
       Object.keys(noteTrackChilds).forEach((noteTrackId) => {
         midiNotesArr.push(createPortal(noteTrackChilds[noteTrackId], props.noteTracksRef.current!.children.namedItem(noteTrackId)!))